refactor: relocate sload benchmark
 - add test_sload file
 - move sload related benchmark
